// client/src/types/common/common.type.ts
import { AxiosError } from "axios";

export type ApiError = AxiosError<{ message?: string }>;

export interface ConfirmationModalProps {
  isOpen: boolean;
  onClose: () => void;
  onConfirm: () => void;
  title: string;
  message: string;
  confirmText?: string;
  cancelText?: string;
  type?: "danger" | "warning" | "info";
}